+++ src/PhpWord/Writer/RTF/Element/AbstractElement.php
@@ -18,9 +18,13 @@
 namespace PhpOffice\PhpWord\Writer\RTF\Element;
 
 use PhpOffice\Common\Text as CommonText;
+use PhpOffice\PhpWord\Element\AbstractElement as Element;
+use PhpOffice\PhpWord\Escaper\Rtf;
+use PhpOffice\PhpWord\Settings;
 use PhpOffice\PhpWord\Style;
 use PhpOffice\PhpWord\Style\Font as FontStyle;
 use PhpOffice\PhpWord\Style\Paragraph as ParagraphStyle;
+use PhpOffice\PhpWord\Writer\AbstractWriter;
 use PhpOffice\PhpWord\Writer\HTML\Element\AbstractElement as HTMLAbstractElement;
 use PhpOffice\PhpWord\Writer\RTF\Style\Font as FontStyleWriter;
 use PhpOffice\PhpWord\Writer\RTF\Style\Paragraph as ParagraphStyleWriter;
@@ -45,6 +49,13 @@ abstract class AbstractElement extends HTMLAbstractElement
      * @var \PhpOffice\PhpWord\Style\Paragraph
      */
     private $paragraphStyle;
+
+    public function __construct(AbstractWriter $parentWriter, Element $element, $withoutP)
+    {
+        parent::__construct($parentWriter, $element, $withoutP);
+
+        $this->escaper = new Rtf();
+    }
 
     /**
      * Get font and paragraph styles.
@@ -112,7 +123,11 @@ abstract class AbstractElement extends HTMLAbstractElement
      */
     protected function writeText($text)
     {
-        return CommonText::toUnicode($text);
+        if (Settings::isOutputEscapingEnabled()) {
+            return $this->escaper->escape($text);
+        } else {
+            return CommonText::toUnicode($text);
+        }
     }
 
     /**